+++ packages/gallery/src/components/item/itemView.js
@@ -1,5 +1,5 @@
 import React from 'react';
-import { GALLERY_CONSTS, featureManager, window, utils, isEditMode, isSiteMode, isSEOMode } from 'pro-gallery-lib';
+import { GALLERY_CONSTS, featureManager, window, utils, isEditMode, isPreviewMode, isSiteMode, isSEOMode } from 'pro-gallery-lib';
 import ImageItem from './imageItem.js';
 import TextItem from './textItem.js';
 import ItemHover from './itemHover.js';
@@ -566,7 +566,8 @@ class ItemView extends GalleryComponent {
         : 'pointer'
     };
 
-    const layoutStyles = settings.avoidInlineStyles ? {} : {
+    const {avoidInlineStyles = true} = settings;
+    const layoutStyles = avoidInlineStyles ? {} : {
       top: offset.top,
       left: styleParams.isRTL ? 'auto' : offset.left,
       right: !styleParams.isRTL ? 'auto' : offset.left,
@@ -574,7 +575,7 @@ class ItemView extends GalleryComponent {
       height: style.height + style.infoHeight,
     };
 
-    const transitionStyles = (this.state.loaded && isEditMode()) ? {
+    const transitionStyles = (this.state.loaded && (isEditMode() || isPreviewMode())) ? {
       transition: 'all .4s ease',
       transitionProperty: 'top, left, width, height'
     } : {