+++ packages/gallery/src/versionLogger.js
@@ -1,1 +1,1 @@
-try {console.debug('Pro Gallery Version: 2.1.26'); } catch (e) {}
+try {console.debug('Pro Gallery Version: 2.1.30'); } catch (e) {}